Generalize Galaxy Item selection field and reuse in visualizations
