+++ client/src/components/Markdown/services.js
@@ -2,8 +2,6 @@ import axios from "axios";
 import { getAppRoot } from "onload/loadConfig";
 import { rethrowSimple } from "utils/simple-error";
 
-import { GalaxyApi } from "@/api";
-
 export async function copyCollection(hdcaId, historyId) {
     const url = `${getAppRoot()}api/histories/${historyId}/contents/dataset_collections`;
     const payload = {
