+++ client/src/components/SelectionField/services.js
@@ -1,5 +1,3 @@
-import axios from "axios";
-import { getAppRoot } from "onload/loadConfig";
 import { rethrowSimple } from "utils/simple-error";
 
 import { GalaxyApi } from "@/api";
